LayoutTests/fast/events/ios/submit-form-target-blank-using-return-key-expected.txt

Tests that pressing the Return key in a text field with an associated form implicitly submits the form. To run this test by hand, focus the text field below and press the Return key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.key is "Enter"
PASS successfullyParsed is true

TEST COMPLETE
